Port to Python3
Tested with Python 3.5.0.

The port mainly consists in handling correctly byte-string and
unicode strings. File buffering is also handled a bit differently,
and a race condition on player termination appeared in my PY3
environment but on in PY2.

+++ orochi/player.py
@@ -216,6 +216,12 @@ class MPlayer(object):
         """Shut down mplayer and replace the reference to the async process
         with a dummy instance that raises a :class:`RuntimeError` on any method
         call."""
+
+        if isinstance(self.p, DeadMPlayer):
+            # self.terminate may be call explicitely *and* implicitely,
+            # through self.__del__.
+            return
+
         if hasattr(self, 'p') and hasattr(self.p, 'terminate'):
             self._stop_background_thread()
             self.p.terminate()
